api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
pydantic==2.9.0
python-dotenv==1.0.0
paho-mqtt==2.1.0
httpx==0.27.0